+++ src/const.ts
@@ -1,4 +1,5 @@
-import { IChartGridStyle, IChartGridStyleInput } from './types';
+import { AxisType } from './layout/axis/axisTypes';
+import { IChartGridStyle, IChartGridStyleInput, Alignment2D } from './types';
 import { Colors } from './utils/colors';
 
 export const kRefLayoutReuseID = 'ref';
@@ -20,3 +21,12 @@ export const kChartGridStyleDarkDefaults = {
     majorGridLineColor: Colors.grey700,
     minorGridLineColor: Colors.grey900,
 } as IChartGridStyle;
+
+export const kDefaultAxisLabelAlignments: {
+    [A in AxisType]: Alignment2D;
+} = {
+    topAxis: { x: 'center', y: 'bottom' },
+    bottomAxis: { x: 'center', y: 'top' },
+    leftAxis: { x: 'right', y: 'center' },
+    rightAxis: { x: 'left', y: 'center' },
+};